pyproject.toml
[project]
name = "audio-chapters-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "click>=8.3.1",
    "eyed3>=0.9.9",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
